% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocod-auth.R
\name{gio_auth}
\alias{gio_auth}
\title{Get or set GEOCODIO_API_KEY value}
\usage{
gio_auth(force = FALSE)
}
\arguments{
\item{force}{force setting a new geocod.io API key for the
current environment?}
}
\value{
the geocod.io API key
}
\description{
The API wrapper functions in this package all rely on a geocod.io API
key residing in the environment variables \code{GEOCODIO_API_KEY}. The easiest way to
accomplish this is to set it in the \code{.Renviron} file in your home directory.
}
