% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocod-batch.R
\name{gio_batch_geocode}
\alias{gio_batch_geocode}
\title{Geocode up to 10,000 addresses}
\usage{
gio_batch_geocode(addresses, api_key = gio_auth())
}
\arguments{
\item{addresses}{character vector of addresses (10K max)}

\item{api_key}{\code{geocod.io} API key}
}
\description{
Geocode up to 10,000 addresses
}
\examples{
gio_batch_geocode(c(
  "1109 N Highland St, Arlington VA",
  "525 University Ave, Toronto, ON, Canada",
  "4410 S Highway 17 92, Casselberry FL",
  "15000 NE 24th Street, Redmond WA",
  "17015 Walnut Grove Drive, Morgan Hill CA"
))
}
