% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocod-reverse-batch.R
\name{gio_batch_reverse}
\alias{gio_batch_reverse}
\title{Reverse geocode up to 10,000 coordinates}
\usage{
gio_batch_reverse(coordinates, fields, api_key = gio_auth())
}
\arguments{
\item{coordinates}{data frame with coordinates in \code{lat} and \code{lon} columns}

\item{fields}{vector of additional fields to return with query results. Note that these
count as extra lookups and impact your dailu quota/costs. See \href{https://geocod.io/docs/#fields}{the official documentation}
for more information on costs/pricing. Can be \code{cd}, \code{cd113}, \code{cd114}, or \code{cd115} for
current or historical Congressional districts (U.S.); \code{stateleg} for State Legislative District (House & Senate, U.S.);
\code{school} forSchool District (elementary/secondary or unified, U.S.); \code{census} for
Census Block/Tract & FIPS codes (U.S.) or \code{timezone} for timezone.}

\item{api_key}{\code{geocod.io} API key}
}
\value{
\code{tibble}
}
\description{
If you have several addresses that you need to reverse geocode, batch rev-geocoding is a much
faster option since it removes the overhead of having to perform multiple HTTP requests.
}
\examples{
data.frame(
  lat = c(35.9746000, 32.8793700, 33.8337100, 35.4171240),
  lon = c(-77.9658000, -96.6303900, -117.8362320, -80.6784760)
) -> to_code
gio_batch_reverse(to_code)
gio_batch_reverse(to_code, fields=c("census", "stateleg"))
}
\references{
\href{https://geocod.io/docs/}{Official Geocodio API documentation}
}
