% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocod-one.R
\name{gio_geocode}
\alias{gio_geocode}
\title{Geocode a single address}
\usage{
gio_geocode(address, api_key = gio_auth())
}
\arguments{
\item{address}{address to geocode}

\item{api_key}{\code{geocod.io} API key}
}
\description{
Geocode a single address
}
\examples{
gio_geocode("1109 N Highland St, Arlington, VA")
}
