% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocod-one.R
\name{gio_geocode_components}
\alias{gio_geocode_components}
\title{Geocode a single address}
\usage{
gio_geocode_components(street, city, state, postal_code, country, fields,
  api_key = gio_auth())
}
\arguments{
\item{street, city, state, postal_code, country}{address components}

\item{fields}{vector of additional fields to return with query results. Note that these
count as extra lookups and impact your dailu quota/costs. See \href{https://geocod.io/docs/#fields}{the official documentation}
for more information on costs/pricing. Can be \code{cd}, \code{cd113}, \code{cd114}, or \code{cd115} for
current or historical Congressional districts (U.S.); \code{stateleg} for State Legislative District (House & Senate, U.S.);
\code{school} forSchool District (elementary/secondary or unified, U.S.); \code{census} for
Census Block/Tract & FIPS codes (U.S.) or \code{timezone} for timezone.}

\item{api_key}{\code{geocod.io} API key}
}
\description{
\strong{You can also geocode intersections.} Just specify the two streets that you want to
geocode in your query.
}
\examples{
gio_geocode_components("1109 N Highland St", "Arlington", "VA")
gio_geocode_components("1109 N Highland St", "Arlington", "VA",
                       fields=c("census", "stateleg"))
}
