% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocod-reverse.R
\name{gio_reverse}
\alias{gio_reverse}
\title{Reverse geocode a single lat/lon pair}
\usage{
gio_reverse(lat, lon, api_key = gio_auth())
}
\arguments{
\item{lat, lon}{lat/lon pair to reverse geocode}

\item{api_key}{\code{geocod.io} API key}
}
\description{
Reverse geocode a single lat/lon pair
}
\examples{
gio_reverse(38.9002898, -76.9990361)
}
