% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocod-reverse-batch.R
\name{gio_batch_reverse}
\alias{gio_batch_reverse}
\title{Reverse geocode up to 10,000 coordinates}
\usage{
gio_batch_reverse(coordinates, api_key = gio_auth())
}
\arguments{
\item{coordinates}{data frame of coordinates with \code{lat} and \code{lon} columns}

\item{api_key}{\code{geocod.io} API key}
}
\description{
Reverse geocode up to 10,000 coordinates
}
\examples{
data.frame(
  lat = c(35.9746000, 32.8793700, 33.8337100, 35.4171240),
  lon = c(-77.9658000, -96.6303900, -117.8362320, -80.6784760)
) -> to_code
gio_batch_reverse(to_code)
}
