% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocod-one.R
\name{gio_geocode_components}
\alias{gio_geocode_components}
\title{Geocode a single address}
\usage{
gio_geocode_components(street, city, state, postal_code, country,
  api_key = gio_auth())
}
\arguments{
\item{street, city, state, postal_code, country}{address components}

\item{api_key}{\code{geocod.io} API key}
}
\description{
Geocode a single address
}
\examples{
gio_geocode_components("1109 N Highland St", "Arlington", "VA")
}
